用户微服务开发-水平权限攻击问题修复

+++ xdclass-user-service/src/main/java/net/xdclass/service/impl/AddressServiceImpl.java
@@ -28,7 +28,8 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public AddressVO detail(Long id) {
-        AddressDO addressDO = addressMapper.selectOne(new QueryWrapper<AddressDO>().eq("id", id));
+        LoginUser loginUser = LoginInterceptor.threadLocal.get();
+        AddressDO addressDO = addressMapper.selectOne(new QueryWrapper<AddressDO>().eq("id", id).eq("user_id", loginUser.getId()));
         if (addressDO == null) {
             return null;
         }
@@ -78,7 +79,8 @@ public class AddressServiceImpl implements AddressService {
      */
     @Override
     public int del(Long addressId) {
-        int rows = addressMapper.delete(new QueryWrapper<AddressDO>().eq("id", addressId));
+        LoginUser loginUser = LoginInterceptor.threadLocal.get();
+        int rows = addressMapper.delete(new QueryWrapper<AddressDO>().eq("id", addressId).eq("user_id", loginUser.getId()));
         return rows;
     }
